Make a separate extension for Combine.
